requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch

fastapi
uvicorn
pymupdf
sentence-transformers
faiss-cpu
pydantic
python-multipart
numpy
openai
streamlit
requests